Implement the list and info commands.
The parts of the Bintray REST API we need are now accessible anonymously, so
I have implemented the list and info commands based on that API. I have also
done some refactoring to keep the REST access in one place and so that it can
be easily replaced later if necessary.

These commands also have integration tests that are taking advantage of the
[Betamax](http://freeside.co/betamax/) library I just incorporated into the
build.