Use 'if' in place of 'try' for flow control.
The app now checks whether the config file exists before trying to load it.
Previously, it attempted the load and then caught the `FileNotFoundException`,
which isn't ideal for an expected condition, i.e. the config file not existing.

The static final fields in `LazybonesMain` have also been capitalised so that
they are more recognisably constants.