Add process timeout for functional tests.
If a command never returned, the functional tests would just hang. This change
puts a 7s timeout on the external processes that the functional tests launch.
It also prints out the command line used to launch the process to make it
easier to identify the command output in the test reports.