Allow templates that are only in the local cache.
Fixes issue #39. This required a pretty significant refactoring to ensure that
Lazybones only connects to the internet when absolutely necessary. But, the
`CreateCommand` class was overdue a refactoring.

The most significant change is that error handling is now done through
exceptions. This allowed the necessary restructing of the command's other
code and should make it easy to handle connection issues properly.