Handle packages with no published versions.
Fixes issue #59. It's possible for a package to be on Bintray without any
associated versions being published. This scenario was causing an exception
with no useful information for the end user.

The `info` and `create` commands now produces a simple error message saying
that no versions have been published for the package.