Fix handling of empty user config file.
Lazybones was assuming that if a config file existed, it specified the list of
Bintray repositories to search for templates. This wasn't the case and most of
the commands errored out due to there being no repositories configured.

Lazybones now sets a default repository list if the user hasn't set one him- or
herself, regardless of whether the config file exists or not.