Add git repository initialisation to `create` command.
Implements issue #53. The user can now pass a `--with-git` option to the
`create` command which results Lazybones creating a new git repository in the
new project. Lazybones also adds all the files in the directory and commits
them.

Template authors can control which files are actually committed through a new
`scmExclusions(String... patterns)` method in the post-install script. This
will copy the given exclusion patterns into a .gitignore file in the root of
the new project.