Improve configuration processing.
Several of the static constants lurking in the code are now settings in a new
defaultConfig.groovy file. These defaults can be overridden by the user's own
configuration file (whose location is itself configurable). These settings can
in turn be overridden by system properties of the form `lazybones.<setting>`.
So `config.file` can be overridden by `lazybones.config.file`.

The integration tests are also set up to run with a particular user config file
on developer machines. The default configuration is used on the CI server.