Add support for sub-templates.
Implementation of issue #86. Sub-templates are just like normal project
templates, except they are not published separately. Instead, they are included
directly in a project template package.

The user can run the new `generate` command to activate a sub-template whose
job it is to actually do something, such as create new files in the project.
The user can also pass in parameters if the sub-template uses the `ask()`
method.

The Gradle plugin for managing Lazybones templates now allows you to define
relationships between main templates and sub-templates. The sub-templates are
packaged separately and then included into the main template package zip.

You create sub-templates as normal templates except that the directory name
must begin with 'subtmpl-'. This ensures that they aren't included in the
`packageAllTemplates`, `installAllTemplates` and `publishAllTemplates` as they
shouldn't be published independently. You then use the Lazybones extension to
declare that a template includes some set of sub-templates. The plugin does
the rest.