Merge stdout and stderr for functional tests.
It was difficult to verify the order in which Lazybones printed stuff to the
console, particularly with stacktraces, because two separate threads were
consuming the output from stderr and stdout. This often left stacktraces and
normal output interleaved.

I now use `ProcessBuilder` explicitly and configure it to merge stdout and
stderr, which eliminates the interleaving problem.