Support creating packages in Bintray.
Implements issue #112 and I think it fixes issue #111 as well.

The Gradle plugin now automatically creates a package in Bintray if it doesn't
already exist. For this to work, you have to specify a `licenses` property in
the `lazybones` block.

You can also now specify just a Bintray repository name rather than a complete
repository URL. This takes the form "<user>/<repo>".