Add runtime configuration manipulation.
Implements issue #118. This feature adds several commands to Lazybones:

    config set <option> <value1> ...
    config add <option> <value1> ...
    config clear <option>
    config show [--all] <option>
    config list

These commands allow users to access the saved Lazybones configuration without
diving into the standard user configuration file, ~/.lazybones/config.groovy.
In addition, it makes instructions for setting up additional template
repositories trivially easy:

    "Just run the command 'lazybones config add bintrayRepositories ...'"

The core of the implementation is a standalone `Configuration` class that does
most of the work of maintaining multiple configuration sources and ensuring
that users are aware when settings are overridden.

The class has an instance-level API that is agnostic of Lazbones, making it
reusable in other projects. The static properties and methods are definitely
Lazybones specific though.