Allow numbers in project template names (Gradle plugin).
Fixes issue #123. The `package...` tasks weren't working for project templates
that had numbers in their names. This change allows people to use things like
Neo4j and EasyAs123, which get converted to neo4j and easy-as123 respectively.

If the name conversion doesn't work for you, there is also the option to run
the package task in hyphenated for, e.g.

    gradle packageTemplate-easy-as-123

This works as long as the first character after `packageTemplate` is a hyphen.