Allow users to set file permissions within packages.
Implements issue #124. The Gradle plugin's extension now allows users to
specify file modes on files and directories in a package. This means that the
package will contain the correct file permissions regardless of which platform
the package is created on (looking at you Windows).

The plugin also now allows users to specify exclusions and file modes on a
per-package basis using the template convention (originally created to support
sub-templates). When specified, the plugin ignores the global exclusions and
file modes when the template is packaged.