Defer setting upload task's artifactFile.
The plugin currently fails to publish Lazybones templates because the upload
task requires the location of the artifact file during the configuration phase.
It also requires the file to exist. Since the location of the file isn't known
until execution time now, the build throws an exception.

This change initialises the upload task's `artifactFile` property at execution
time, while also removing the property's `@InputFile` annotation. It seems that
the incremental build support works during the configuration phase.

The `artifactUrlPath` property is also now set at execution time because it
includes the package version, which is not known until then.