Evaluate template version at config time in plugin.
The fix for issue #126 involved evaluating the template version at execution
time when packaging a template. Unfortunately, this broke various aspects of
the packaging, particularly the embedding of sub-templates.

This fixes issue #131 by reverting back to evaluating the version at config
time again. In order to satisfy the needs of #126, users can now specify the
template version in the Lazybones build convention:

    lazybones {
        template("std-java") {
            version = "1.0.1"
        }
    }

Note that the version is a template-specific property.