Fix integration tests on Java 8.
Running the integration tests on the Java 8 VM resulted in lots of class
loading exceptions related to HashMap. This was because the integration tests
were running with version 2.0.5 of Groovy, which was pulled in by Spock. The
tests now run with the same version of Groovy as the app and other parts of the
build.