Add warning for template dirs with non-conventional names.
Helps with issue #138. The Gradle plugin previously caused a mysterious error
when it tried to process a template directory not matching the lowercase
hyphenated form required by the conventions.

This fix performs a check on all template directories to see whether they can
be automatically added to `packageAllTemplates` et al. Those that can't trigger
a warning message and are basically ignored. Users can still manually add them
as dependencies of the `pckageAllTemplates` task though.