Force `ask()` default value to String.
Fixes issue #135. Both incoming values (from the user at the command line) and
outgoing values (in the saved parent parameters) are based on strings, so the
`ask()` method now requires its default values to be strings as well. This means
that the template developer needs to manage any necessary conversions to other
types.