Add info to README.md mentioning relationship to Maven archetypes and Yeoman
It's helpful to add information to the overview that mentions how this tool is similar to other tools in related domains.  In particular, the Lazybones project does something very similar to what Maven does with archetypes, and what the Yeoman project does for JavaScript-based web applications. The added paragraph hopefully puts some light on that.