Allow values of any type for system properties.
Fixes issue #155. It was possible to set the proxy port to an integer in the
Lazybones configuration file, but this caused an error because only strings
were allowed for system properties. The validation now allows any type and it's
up to the user to ensure that the string versions of the values are appropriate.