Add VCS URL for Bintray upload.
Fixes issue #154. Bintray requires a VCS URL for OSS packages, which the plugin
didn't allow the user to set. You can now set a `vcsUrl` property in the
Lazybones configuration block. Note that this is not required if the package
already exists or you are using paid (i.e. non-OSS) hosting.