Add --subs option to `list` command.
Further work to #147. There is now a `--subs` option on the `list` command that
only displays the subtemplates installed in the current Lazybones-created
project. The command errors out if there is no .lazybones directory in the
current working directory.

I also modified the `--cached` option to group templates by name, so that all
available versions are listed as one entry.