Upgrade to Groovy WSLite to 1.1.2.
This version has builtin support for the standard JVM proxy properties, so there
is no longer any need to manually configure the proxy. I've also fixed a nasty
hack that disabled SSL certificate verification if a proxy was being used. This
now only happens during integration tests, or more specifically if an
'integration.test' system property is set to the string `"true"`.